Renamed all plugins to be less messy
