+++ src/plugins/filter.ts
@@ -1,4 +1,4 @@
-import { BasePlugin } from './base.plugin';
+import { BasePlugin } from './base';
 import * as Discord from 'discord.js';
 import * as fs from 'fs';
 import * as pluralize from 'pluralize';
